Homemade physics

+++ src/Graphics/CanvasInterface.java
@@ -5,6 +5,6 @@ import java.awt.Image;
 public interface CanvasInterface {   
    public ColorMode  setGraphicsMode(ColorMode newMode);
    public void       drawImage(ImageItem imageItem);
-   public void       drawImage(Image image, float x, float y, float z, float width, float height);
+   public void       drawImage(Image image, double x, double y, double z, double width, double height);
    public void       draw();
 }
